requirements.txt
selenium
pandas
python-dotenv
apscheduler
click
webdriver-manager
psycopg2-binary
cryptography
